Add zshenv file.
The zshenv file immediately defines ZDOTDIR and exports a path for use
with the Haskell for OS X app.

Restructured where zsh files live.  Now we create a .zsh dir in the home
dir and symlink from the dotfiles repo into this.  The .zsh file
contains everything related to zsh, theoretically, such as the command
history.