Implement basic plugin manager.
A basic plugin manager is implemented in `plugins`. Currently the
functionality is mostly programatic. To use, simply source
`plugins/init.zsh` in `.zshrc`.

The `plugins/bin/plug.zsh` file exposes some basic functions to manage
plugins, modeled after `zplug`. The exposed commands are `plugins/bin/install`
and `plugins/bin/cache`.

The `cache` command creates a static `plugins/cache.zsh` file with
source statements for each plugin listed in `plugins/plugins.zsh`. This
file is sourced or created in `plugins/init.zsh`.

Overall, we see roughly a 200-300ms startup time improvement over zsh,
because once the cache file is created, nothing else needs to be loaded.

Right now the only feature is the ability to specify a custom file to
source in a plugin. For example, the `z` plugin is utilized by sourcing
`z.sh`. In the future we might support more of zplug's features, but if
we end up needing them we will likely switch to that package manager to
avoid future maintenance costs.

The `fzf` zsh key bindings and completions are now considered plugins.
The binary will be installed via `nix`. It's possible this is a bad
idea, as the repo fetched for zsh might not coincide with the binary
provided by `nix`, but hopefully the completions and key bindings are
relatively static.